requirements.txt
pyyaml
requests
python-dotenv
pytest